feat: include local tunnel work
Update modules to be typescript instead of javascript
Update output package json command
Update build command to clean dist before building
